requirements.txt
bcrypt==4.2.0
blinker==1.8.2
certifi==2024.8.30
charset-normalizer==3.3.2
click==8.1.7
Flask==3.0.3
idna==3.10
importlib_metadata==8.5.0
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
PyJWT==2.9.0
pymssql==2.3.1
pyodbc==5.1.0
python-dotenv==1.0.1
pytz==2024.2
requests==2.32.3
schedule==1.2.2
urllib3==2.2.3
Werkzeug==3.0.4
zipp==3.20.2
